requirements.txt
PyPDF2>=3.0.0
pdfplumber>=0.7.0
pdfminer.six>=20221105
python-docx>=0.8.11
docx2txt>=0.8